+++ BackEnd/Services/JwtHandlerService.cs
@@ -12,22 +12,18 @@ namespace TheMenu.BackEnd.Services
 {
     public class JwtHandlerService
     {
-		private readonly IConfiguration _configuration;
-        private readonly IConfigurationSection _jwtSettings;
         private readonly EnvironmentSpecificSettings _settings;
         private readonly UserManager<TheMenuBackEndUser> _userManager;
 
         public JwtHandlerService(IConfiguration configuration, UserManager<TheMenuBackEndUser> userManager)
         {
             _userManager = userManager;
-            _configuration = configuration;
-            _jwtSettings = _configuration.GetSection("JwtSettings");
             _settings = configuration.Get<EnvironmentSpecificSettings>();
         }
 
         private SigningCredentials GetSigningCredentials()
         {
-            var key = Encoding.UTF8.GetBytes(_jwtSettings.GetSection("securityKey").Value);
+            var key = Encoding.UTF8.GetBytes(_settings.JwtSecretKey);
             var secret = new SymmetricSecurityKey(key);
 
             return new SigningCredentials(secret, SecurityAlgorithms.HmacSha256);
@@ -52,10 +48,10 @@ namespace TheMenu.BackEnd.Services
         private JwtSecurityToken GenerateTokenOptions(SigningCredentials signingCredentials, List<Claim> claims)
         {
             var tokenOptions = new JwtSecurityToken(
-                issuer: _jwtSettings.GetSection("validIssuer").Value,
-                audience: _jwtSettings.GetSection("validAudience").Value,
+                issuer: _settings.JwtValidIssuer,
+                audience: _settings.JwtValidAudience,
                 claims: claims,
-                expires: DateTime.Now.AddMinutes(Convert.ToDouble(_jwtSettings.GetSection("expiryInMinutes").Value)),
+                expires: DateTime.Now.AddMinutes(5),
                 signingCredentials: signingCredentials);
 
             return tokenOptions;