+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Waiting for EF tooling for .NET 6 - https://gist.github.com/davidfowl/0e0372c3c1
 - ApplicationInsights - backend
 - Diagnose / ping
 - ASP.NET Identity + EntityFramework Core Migrations [Scaffold Identity in ASP.NET Core projects], [Tutorial: Get started with EF Core in an ASP.NET MVC web app]
+- ASP.NET Core Authentication with JWT and Angular - [ASP.NET Core Authentication with JWT and Angular – Part 1]
 - [Angular Routing]
 
 ## [v0.0.1] (2021-09-23)
@@ -61,4 +62,5 @@ Waiting for EF tooling for .NET 6 - https://gist.github.com/davidfowl/0e0372c3c1
 [ASP.NET Core Identity]: https://docs.microsoft.com/en-us/aspnet/core/security/authentication/scaffold-identity?view=aspnetcore-6.0&tabs=visual-studio#scaffold-identity-into-a-razor-project-without-existing-authorization
 [Scaffold Identity in ASP.NET Core projects]: https://docs.microsoft.com/en-us/aspnet/core/security/authentication/scaffold-identity?view=aspnetcore-6.0&tabs=visual-studio#scaffold-identity-into-a-razor-project-without-existing-authorization
 [Run EF Core Migrations in Azure DevOps]: https://dotnetthoughts.net/run-ef-core-migrations-in-azure-devops/
+[ASP.NET Core Authentication with JWT and Angular – Part 1]: https://code-maze.com/authentication-aspnetcore-jwt-1/
 [Angular Routing]: https://angular.io/guide/router